Componentize panel
