--- amazon-assistant/panel/src/Panel.css
@@ -1,214 +1,0 @@
-/* :root vars from https://github.com/biancadanforth/contextual-feature-recommendation/blob/master/amazon-assistant/notification-bar/notificationBar.css */
-:root {
-  --primary-color: #fdfdfd;
-  --primary-text-color: #6f6f6f;
-  --link-text-color: #0094fb;
-  --primary-border-color: #b7b7b7;
-  --standard-box-shadow: 0 10px 20px rgba(50,50,93,.1),0 3px 7px rgba(50,50,93,.15),0 1px 3px rgba(0,0,0,.1);
-  --dropdown-box-shadow: 0 50px 100px rgba(50, 50, 93, .1), 0 15px 35px rgba(50, 50, 93, .15), 0 5px 15px rgba(0, 0, 0, .1);
-  --accept-button-background-color: #0996f8;
-  --accept-button-hover-background-color: #00329a;
-  --reject-button-background-color: #f0f0f0;
-  --reject-button-hover-background-color: #e1e1e2;
-}
-
-* {
-  box-sizing: border-box;
-}
-
-body {
-  font-size: .85em;
-}
-
-p {
-  margin: 0;
-}
-
-a {
-  color: #5fbdf8;
-}
-
-.Panel {
-  max-width:400px;
-  background-color: var(--primary-color);
-  border: 1px solid var(--primary-border-color);
-  box-shadow: var(--standard-box-shadow);
-}
-
-.section-grid-wrapper {
-  display: grid;
-  grid-template-columns: 30px auto;
-  grid-template-rows: 30px auto;
-  align-items: center;
-  grid-gap: 10px;
-  padding: 20px;
-}
-
-/* ----- Top section  ----- */
-
-.section-top {
-  display: grid;
-  grid-gap: 10px;
-  grid-template-columns: 30px auto;
-  grid-column: 1 / 3;
-}
-
-.icon {
-  width: 30px;
-  height: 30px;
-}
-
-.header {
-  grid-column-start: 2;
-}
-
-.rationale {
-  grid-column-start: 3;
-  align-self: flex-start;
-  width: 15px;
-}
-
-/* ----- Middle section  ----- */
-
-.section-middle {
-  grid-column-start: 2;
-  grid-row-start: 2;
-  display: flex;
-  align-items: center;
-}
-
-.section-middle div {
-  margin-left: 15px;
-}
-
-.users {
-  display: inline-block;
-  margin-left: 5px;
-  color: #a3a3a3;
-}
-
-
-/* ----- Bottom section  ----- */
-
-.section-bottom {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-}
-
-
-/* <Button /> */
-
-.button {
-  font-size: inherit;
-  padding: 10px;
-  cursor: pointer;
-  color: var(--primary-text-color);
-}
-
-.primary-button {
-  color: white;
-  background-color: var(--accept-button-background-color);
-  border-color: #3085d2;
-}
-
-.primary-button:hover {
-  background-color: var(--accept-button-hover-background-color);
-}
-
-.secondary-button-wrapper {
-  display: grid;
-  grid-template-columns: 1fr auto;
-}
-
-.secondary-button {
-  background-color: var(--reject-button-background-color);
-  border-right-color: rgba(0, 0, 0, .2);
-}
-
-.secondary-button:hover, #secondary-button-show-dropdown:hover {
-  background-color: var(--reject-button-hover-background-color);
-}
-
-#secondary-button-show-dropdown {
-  position: relative;
-  margin: 0;
-  /* From: resource://gre-resources/forms.css in Firefox URL bar*/
-  border: 2px outset ThreeDLightShadow;
-  border-left-width: 0;
-  align-self: stretch;
-  background-color: var(--reject-button-background-color);
-}
-
-#secondary-button-show-dropdown:active {
-  border-style: inset;
-}
-
-#secondary-button-show-dropdown::after {
-  content: "";
-  display: inline-block;
-  background-repeat: no-repeat;
-  background-position: center;
-  width: 10px;
-  height: 10px;
-  background-size: contain;
-  background-image: url(down-arrow.svg);
-}
-
-/* <Dropdown /> */
-
-
-.dropdown {
-  position: absolute;
-  top: 100%;
-  left: 50%;
-  transform: translate(-50%, 15px);
-  min-width: 200px;
-  visibility: visible;
-  margin: 0 auto;
-  opacity: 1;
-  transition: opacity 0.5s;
-  flex-flow: column wrap;
-  padding: 0;
-  margin: 0;
-  text-align: start;
-  list-style: none;
-  box-shadow: var(--dropdown-box-shadow);
-}
-
-.dropdown::before {
-  position: absolute;
-  left: 50%;
-  top: 0%;
-  transform: translate(-50%, -10px);
-  width: 0;
-  height: 0;
-  content:'';
-  border-left: 10px solid transparent;
-  border-right: 10px solid transparent;
-  border-bottom: 10px solid var(--reject-button-background-color);
-}
-
-.dropdown-item {
-  padding: 10px;
-  border-bottom: 1px solid var(--primary-border-color);
-  background-color: var(--reject-button-background-color);
-  transition: background-color 0.5s;
-}
-
-.dropdown-item:last-child {
-  border: 0;
-}
-
-.dropdown-item:hover {
-  background-color: var(--reject-button-hover-background-color);
-}
-
-.hidden {
-  visibility: hidden;
-  opacity: 0;
-  transition: visibility 0s linear 0.5s, opacity 0.5s;
-}
-
-.button.hidden {
-  display: none;
-}
