+++ amazon-assistant/panel/src/css/Panel.css
@@ -1,0 +1,96 @@
+/* :root vars from https://github.com/biancadanforth/contextual-feature-recommendation/blob/master/amazon-assistant/notification-bar/notificationBar.css */
+:root {
+  --primary-color: #fdfdfd;
+  --primary-text-color: #6f6f6f;
+  --link-text-color: #0094fb;
+  --primary-border-color: #b7b7b7;
+  --standard-box-shadow: 0 10px 20px rgba(50,50,93,.1),0 3px 7px rgba(50,50,93,.15),0 1px 3px rgba(0,0,0,.1);
+  --dropdown-box-shadow: 0 50px 100px rgba(50, 50, 93, .1), 0 15px 35px rgba(50, 50, 93, .15), 0 5px 15px rgba(0, 0, 0, .1);
+  --accept-button-background-color: #0996f8;
+  --accept-button-hover-background-color: #00329a;
+  --reject-button-background-color: #f0f0f0;
+  --reject-button-hover-background-color: #e1e1e2;
+}
+
+* {
+  box-sizing: border-box;
+}
+
+body {
+  font-size: .85em;
+}
+
+p {
+  margin: 0;
+}
+
+a {
+  color: #5fbdf8;
+}
+
+.Panel {
+  max-width:400px;
+  background-color: var(--primary-color);
+  border: 1px solid var(--primary-border-color);
+  box-shadow: var(--standard-box-shadow);
+}
+
+.section-grid-wrapper {
+  display: grid;
+  grid-template-columns: 30px auto;
+  grid-template-rows: 30px auto;
+  align-items: center;
+  grid-gap: 10px;
+  padding: 20px;
+}
+
+/* ----- Top section  ----- */
+
+.section-top {
+  display: grid;
+  grid-gap: 10px;
+  grid-template-columns: 30px auto;
+  grid-column: 1 / 3;
+}
+
+.icon {
+  width: 30px;
+  height: 30px;
+}
+
+.header {
+  grid-column-start: 2;
+}
+
+.rationale {
+  grid-column-start: 3;
+  align-self: flex-start;
+  width: 15px;
+}
+
+/* ----- Middle section  ----- */
+
+.section-middle {
+  grid-column-start: 2;
+  grid-row-start: 2;
+  display: flex;
+  align-items: center;
+}
+
+.section-middle div {
+  margin-left: 15px;
+}
+
+.users {
+  display: inline-block;
+  margin-left: 5px;
+  color: #a3a3a3;
+}
+
+
+/* ----- Bottom section  ----- */
+
+.section-bottom {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+}
