+++ amazon-assistant/panel/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import Panel from './Panel';
+import Panel from './Panel.jsx';
 
 ReactDOM.render(<Panel />, document.getElementById('root'));
